+++ todolist/src/App.js
@@ -1,7 +1,6 @@
 import React from "react";
-import logo from "./logo.svg";
 import "./App.css";
-import Todolist from "./Todolist";
+import { Todolist } from "./Todolist";
 
 function App() {
   return (
