+++ todolist/src/Todolist.js
@@ -1,39 +1,37 @@
+// Vendor
 import React from "react";
-import ListTasks from "./ListTasks.js";
 
-function Todolist(props) {
+// Internal
+import { AddTask } from "./AddTask.js";
+import { Task } from "./Task.js";
+
+const Todolist = (props) => {
   const { title } = props;
-  const [taskName, setTaskName] = React.useState("");
-  const listTasksRef = React.useRef();
 
-  const handleSubmit = (event) => {
-    event.preventDefault();
-    if (event.target.elements.addTask.value === "") {
-      alert("You entered an empty string");
-    } else {
-      listTasksRef.current.updateTasks(taskName);
-      setTaskName("");
-    }
+  // Hooks
+  const [tasks, setTasks] = React.useState([]);
+
+  // Handlers
+  const addTask = (taskName) => {
+    setTasks([...tasks, taskName]);
   };
 
-  const handleChange = (event) => {
-    setTaskName(event.target.value);
+  // Markup
+  const renderTasks = () => {
+    const element = tasks.map((task, index) => (
+      <Task texte={task} key={index} />
+    ));
+    return element;
   };
 
   return (
     <div>
       <h2>{title}</h2>
-      <form onSubmit={handleSubmit}>
-        <div>
-          <label htmlFor="addTask">Task:</label>
-          <input value={taskName} onChange={handleChange} id="addTask"></input>
-        </div>
-        <button type="submit">{`Add task ${title}`}</button>
-      </form>
+      <AddTask addTask={addTask} title={title} />
       <br />
-      <ListTasks ref={listTasksRef} />
+      {renderTasks()}
     </div>
   );
-}
+};
 
-export default Todolist;
+export { Todolist };
